Move core/profile/context.h to core/providers/cuda/nvtx_profile_context.h.
